feat: Consolidate CLI into single maintainable file (Phase 2 complete)
Major improvements:
- Merged unified CLI logic from cli_unified.py into cli.py
- Single source of truth for all CLI functionality (787 lines)
- Intelligent routing between unified and legacy commands
- Full backward compatibility with deprecation warnings
- Clean, maintainable architecture ready for production

Benefits:
- No more import complexity or circular dependencies
- Easier to maintain and extend
- Clear migration path for users
- Professional, production-ready code structure

The three workflows (debug/compliance/improve) are now seamlessly
integrated while maintaining support for legacy 20+ flags during
the transition period.
